requirements.txt
selenium-wire == 4.5.5
random-password-generator == 2.2.0
requests == 2.26.0
dynaconf == 3.1.7
Faker == 9.8.2
fake_user_agent == 0.0.12
